add sentry-okhttp module
partially resolves getsentry/sentry-java#1783
