+++ sentry-android-okhttp/api/sentry-android-okhttp.api
@@ -6,8 +6,7 @@ public final class io/sentry/android/okhttp/BuildConfig {
 	public fun <init> ()V
 }
 
-public final class io/sentry/android/okhttp/SentryOkHttpEventListener : okhttp3/EventListener {
-	public static final field Companion Lio/sentry/android/okhttp/SentryOkHttpEventListener$Companion;
+public final class io/sentry/android/okhttp/SentryOkHttpEventListener : io/sentry/okhttp/SentryOkHttpEventListener {
 	public fun <init> ()V
 	public fun <init> (Lio/sentry/IHub;Lkotlin/jvm/functions/Function1;)V
 	public synthetic fun <init> (Lio/sentry/IHub;Lkotlin/jvm/functions/Function1;ILkotlin/jvm/internal/DefaultConstructorMarker;)V
@@ -17,47 +16,16 @@ public final class io/sentry/android/okhttp/SentryOkHttpEventListener : okhttp3/
 	public synthetic fun <init> (Lio/sentry/IHub;Lokhttp3/EventListener;ILkotlin/jvm/internal/DefaultConstructorMarker;)V
 	public fun <init> (Lokhttp3/EventListener$Factory;)V
 	public fun <init> (Lokhttp3/EventListener;)V
-	public fun cacheConditionalHit (Lokhttp3/Call;Lokhttp3/Response;)V
-	public fun cacheHit (Lokhttp3/Call;Lokhttp3/Response;)V
-	public fun cacheMiss (Lokhttp3/Call;)V
-	public fun callEnd (Lokhttp3/Call;)V
-	public fun callFailed (Lokhttp3/Call;Ljava/io/IOException;)V
-	public fun callStart (Lokhttp3/Call;)V
-	public fun canceled (Lokhttp3/Call;)V
-	public fun connectEnd (Lokhttp3/Call;Ljava/net/InetSocketAddress;Ljava/net/Proxy;Lokhttp3/Protocol;)V
-	public fun connectFailed (Lokhttp3/Call;Ljava/net/InetSocketAddress;Ljava/net/Proxy;Lokhttp3/Protocol;Ljava/io/IOException;)V
-	public fun connectStart (Lokhttp3/Call;Ljava/net/InetSocketAddress;Ljava/net/Proxy;)V
-	public fun connectionAcquired (Lokhttp3/Call;Lokhttp3/Connection;)V
-	public fun connectionReleased (Lokhttp3/Call;Lokhttp3/Connection;)V
-	public fun dnsEnd (Lokhttp3/Call;Ljava/lang/String;Ljava/util/List;)V
-	public fun dnsStart (Lokhttp3/Call;Ljava/lang/String;)V
-	public fun proxySelectEnd (Lokhttp3/Call;Lokhttp3/HttpUrl;Ljava/util/List;)V
-	public fun proxySelectStart (Lokhttp3/Call;Lokhttp3/HttpUrl;)V
-	public fun requestBodyEnd (Lokhttp3/Call;J)V
-	public fun requestBodyStart (Lokhttp3/Call;)V
-	public fun requestFailed (Lokhttp3/Call;Ljava/io/IOException;)V
-	public fun requestHeadersEnd (Lokhttp3/Call;Lokhttp3/Request;)V
-	public fun requestHeadersStart (Lokhttp3/Call;)V
-	public fun responseBodyEnd (Lokhttp3/Call;J)V
-	public fun responseBodyStart (Lokhttp3/Call;)V
-	public fun responseFailed (Lokhttp3/Call;Ljava/io/IOException;)V
-	public fun responseHeadersEnd (Lokhttp3/Call;Lokhttp3/Response;)V
-	public fun responseHeadersStart (Lokhttp3/Call;)V
-	public fun satisfactionFailure (Lokhttp3/Call;Lokhttp3/Response;)V
-	public fun secureConnectEnd (Lokhttp3/Call;Lokhttp3/Handshake;)V
-	public fun secureConnectStart (Lokhttp3/Call;)V
 }
 
-public final class io/sentry/android/okhttp/SentryOkHttpEventListener$Companion {
-}
-
-public final class io/sentry/android/okhttp/SentryOkHttpInterceptor : okhttp3/Interceptor {
+public final class io/sentry/android/okhttp/SentryOkHttpInterceptor : io/sentry/okhttp/SentryOkHttpInterceptor, okhttp3/Interceptor {
 	public fun <init> ()V
 	public fun <init> (Lio/sentry/IHub;)V
 	public fun <init> (Lio/sentry/IHub;Lio/sentry/android/okhttp/SentryOkHttpInterceptor$BeforeSpanCallback;ZLjava/util/List;Ljava/util/List;)V
 	public synthetic fun <init> (Lio/sentry/IHub;Lio/sentry/android/okhttp/SentryOkHttpInterceptor$BeforeSpanCallback;ZLjava/util/List;Ljava/util/List;ILkotlin/jvm/internal/DefaultConstructorMarker;)V
+	public fun <init> (Lio/sentry/IHub;Lio/sentry/okhttp/SentryOkHttpInterceptor$BeforeSpanCallback;ZLjava/util/List;Ljava/util/List;)V
+	public synthetic fun <init> (Lio/sentry/IHub;Lio/sentry/okhttp/SentryOkHttpInterceptor$BeforeSpanCallback;ZLjava/util/List;Ljava/util/List;ILkotlin/jvm/internal/DefaultConstructorMarker;)V
 	public fun <init> (Lio/sentry/android/okhttp/SentryOkHttpInterceptor$BeforeSpanCallback;)V
-	public fun intercept (Lokhttp3/Interceptor$Chain;)Lokhttp3/Response;
 }
 
 public abstract interface class io/sentry/android/okhttp/SentryOkHttpInterceptor$BeforeSpanCallback {
