+++ sentry-okhttp/src/main/java/io/sentry/okhttp/SentryOkHttpEvent.kt
@@ -1,4 +1,4 @@
-package io.sentry.android.okhttp
+package io.sentry.okhttp
 
 import io.sentry.Breadcrumb
 import io.sentry.Hint
@@ -9,13 +9,13 @@ import io.sentry.SentryLevel
 import io.sentry.SpanDataConvention
 import io.sentry.SpanStatus
 import io.sentry.TypeCheckHint
-import io.sentry.android.okhttp.SentryOkHttpEventListener.Companion.CONNECTION_EVENT
-import io.sentry.android.okhttp.SentryOkHttpEventListener.Companion.CONNECT_EVENT
-import io.sentry.android.okhttp.SentryOkHttpEventListener.Companion.REQUEST_BODY_EVENT
-import io.sentry.android.okhttp.SentryOkHttpEventListener.Companion.REQUEST_HEADERS_EVENT
-import io.sentry.android.okhttp.SentryOkHttpEventListener.Companion.RESPONSE_BODY_EVENT
-import io.sentry.android.okhttp.SentryOkHttpEventListener.Companion.RESPONSE_HEADERS_EVENT
-import io.sentry.android.okhttp.SentryOkHttpEventListener.Companion.SECURE_CONNECT_EVENT
+import io.sentry.okhttp.SentryOkHttpEventListener.Companion.CONNECTION_EVENT
+import io.sentry.okhttp.SentryOkHttpEventListener.Companion.CONNECT_EVENT
+import io.sentry.okhttp.SentryOkHttpEventListener.Companion.REQUEST_BODY_EVENT
+import io.sentry.okhttp.SentryOkHttpEventListener.Companion.REQUEST_HEADERS_EVENT
+import io.sentry.okhttp.SentryOkHttpEventListener.Companion.RESPONSE_BODY_EVENT
+import io.sentry.okhttp.SentryOkHttpEventListener.Companion.RESPONSE_HEADERS_EVENT
+import io.sentry.okhttp.SentryOkHttpEventListener.Companion.SECURE_CONNECT_EVENT
 import io.sentry.util.Platform
 import io.sentry.util.UrlUtils
 import okhttp3.Request
