+++ sentry-okhttp/src/test/java/io/sentry/okhttp/SentryOkHttpEventListenerTest.kt
@@ -1,4 +1,4 @@
-package io.sentry.android.okhttp
+package io.sentry.okhttp
 
 import io.sentry.BaggageHeader
 import io.sentry.IHub
