+++ sentry-okhttp/src/test/java/io/sentry/okhttp/SentryOkHttpInterceptorTest.kt
@@ -1,6 +1,6 @@
 @file:Suppress("MaxLineLength")
 
-package io.sentry.android.okhttp
+package io.sentry.okhttp
 
 import io.sentry.BaggageHeader
 import io.sentry.Breadcrumb
